+++ src/SMAPI/Framework/ModHelpers/ContentPackHelper.cs
@@ -42,7 +42,15 @@ namespace StardewModdingAPI.Framework.ModHelpers
         public IContentPack CreateFake(string directoryPath)
         {
             string id = Guid.NewGuid().ToString("N");
-            return this.CreateTemporary(directoryPath, id, id, id, id, new SemanticVersion(1, 0, 0));
+            string relativePath = Path.GetRelativePath(Constants.ModsPath, directoryPath);
+            return this.CreateTemporary(
+                directoryPath: directoryPath,
+                id: id,
+                name: $"{this.Mod.DisplayName} (fake content pack: {relativePath})",
+                description: $"A temporary content pack created by the {this.Mod.DisplayName} mod.",
+                author: "???",
+                new SemanticVersion(1, 0, 0)
+            );
         }
 
         /// <inheritdoc />