+++ docs/release-notes.md
@@ -3,6 +3,7 @@
 # Release notes
 ## Upcoming release
 * For mod authors:
+  * Dynamic content packs created via `helper.ContentPacks.CreateTemporary` or `CreateFake` are now listed in the log file.
   * Fixed assets loaded through a fake content pack not working correctly since 3.14.0.
 
 ## 3.14.1